+++ MyPo.Blazor/MyPo.Blazor.Portfolio.App/Services/PortfolioApiClient.FinHub.cs
@@ -85,6 +85,21 @@ public partial class PortfolioApiClient
         return await ReadAndCloseResponseAsync<PortfolioAnalysis>(httpResult, cancellationToken);
     }
 
+    /// <inheritdoc/>
+    public async Task<ApiResp<PortfolioAnalysis>> SpotlightPortfolioPlanAsync(string planId, string authToken, string? baseUrl = default, HttpClient? requestHttpClient = default, CancellationToken cancellationToken = default)
+    {
+        var endpoint = IPortfolioApiClient.API_FINHUB_AI_SPOTLIGHT_PORTFOLIO
+            .Replace("{id}", planId, StringComparison.OrdinalIgnoreCase);
+        using var httpResult = await BuildAndSendRequestAsync(
+            requestHttpClient,
+            HttpMethod.Get, baseUrl, endpoint,
+            authToken,
+            NoData,
+            cancellationToken
+        );
+        return await ReadAndCloseResponseAsync<PortfolioAnalysis>(httpResult, cancellationToken);
+    }
+
     /*----------------------------------------------------------------------*/
 
     /// <inheritdoc />