+++ MyPo.Portfolio.Api/Controllers/FinHubController.AI.cs
@@ -34,15 +34,15 @@ public partial class FinHubController
         var market = Globals.MarketsMap.TryGetValue(portfolio?.Metadata?.DefaultMarketId?.ToUpper() ?? string.Empty, out var m) ? m : null;
 
         /*
-        if current holdings is empty, or >=half of tickers are at 0 allocation ==> call API to build the portfolio
-        otherwise call API to analyze for portfolio
+        if current holdings is empty, or >=half of tickers are at 0 allocation ==> call FinHub API to build the portfolio
+        otherwise call FinHub API to analyze the portfolio
         */
 
-        // Step 1: check portfolio plan's holdings
+        // Step 0: check portfolio plan's holdings
         var countPositiveAllocation = (portfolioPlan.Metadata?.HoldingTickers ?? []).Where(ht => ht.Shares > 0).Count();
         var countEntries = (portfolioPlan.Metadata?.HoldingTickers ?? []).Count;
 
-        // Step 2: make API call
+        // Step 1: make API call
         var finhubResult = countEntries == 0 || (double)countPositiveAllocation / countEntries <= 0.5
             ? await BuildPortfolio(portfolioPlan, market)
             : await AnalyzePortfolio(portfolioPlan, market);
@@ -61,49 +61,112 @@ public partial class FinHubController
         if (!result.LLMError)
         {
             portfolioPlan.Metadata ??= new();
-            portfolioPlan.Metadata.AnalysisRefreshTimestsmp = DateTimeOffset.UtcNow.ToUnixTimeSeconds();
+            portfolioPlan.Metadata.AnalysisRefreshTimestamp = DateTimeOffset.UtcNow.ToUnixTimeSeconds();
             portfolioPlan.Metadata.Analysis = result.Analysis;
             await PortfolioRepository.UpdatePortfolioPlanAsync(portfolioPlan);
         }
         return ResponseOk(result);
     }
 
+    private static List<HoldingTickerReq> BuildHoldingTickersReq(PortfolioPlanEntity plan)
+    {
+        return [.. (plan.Metadata?.HoldingTickers ?? []).Select(ht => new HoldingTickerReq
+        {
+            Ticker = ht.Ticker,
+            TargetAllocation = ht.TargetAllocation,
+            NumShares = ht.Shares,
+            AvgPrice = ht.AveragePrice,
+            MarketPrice = ht.MarketPrice,
+            Tags = ht.Tags,
+        })];
+    }
+
     private async ValueTask<ApiResp<PortfolioAnalysis>> BuildPortfolio(PortfolioPlanEntity plan, MarketDef? market)
     {
+        Logger?.LogInformation("Calling FinHub BuildPortfolio API for portfolio plan '{planId}: {planName}' (market: {market}) with {numHoldings} holdings...", plan.Id, plan.Name, market?.Code ?? "N/A", plan.Metadata?.HoldingTickers?.Count ?? 0);
         var req = new BuildPortfolioReq
         {
             Country = market?.Country ?? "US",
             InvestorTheme = plan.Metadata?.Description,
-            CurrentAllocation = [.. (plan.Metadata?.HoldingTickers?? []).Select(ht => new HoldingTickerReq
-            {
-                Ticker = ht.Ticker,
-                TargetAllocation = ht.TargetAllocation,
-                NumShares = ht.Shares,
-                AvgPrice = ht.AveragePrice,
-                MarketPrice = ht.MarketPrice,
-                Tags = ht.Tags,
-            })]
+            CurrentAllocation = BuildHoldingTickersReq(plan)
         };
         return await FinHubClient.BuildPortfolioAsync(req);
     }
 
     private async ValueTask<ApiResp<PortfolioAnalysis>> AnalyzePortfolio(PortfolioPlanEntity plan, MarketDef? market)
     {
-        var req = new BuildPortfolioReq
+        Logger?.LogInformation("Calling FinHub AnalyzePortfolio API for portfolio plan '{planId}: {planName}' (market: {market}) with {numHoldings} holdings...", plan.Id, plan.Name, market?.Code ?? "N/A", plan.Metadata?.HoldingTickers?.Count ?? 0);
+        var req = new AnalyzePortfolioReq
         {
             Country = market?.Country ?? "US",
             InvestorTheme = plan.Metadata?.Description,
-            CurrentAllocation = [.. (plan.Metadata?.HoldingTickers?? []).Select(ht => new HoldingTickerReq
-            {
-                Ticker = ht.Ticker,
-                TargetAllocation = ht.TargetAllocation,
-                NumShares = ht.Shares,
-                AvgPrice = ht.AveragePrice,
-                MarketPrice = ht.MarketPrice,
-                Tags = ht.Tags,
-            })]
+            CurrentAllocation = BuildHoldingTickersReq(plan)
         };
-        return await FinHubClient.BuildPortfolioAsync(req);
+        return await FinHubClient.AnalyzePortfolioAsync(req);
+    }
+
+    [HttpGet(IPortfolioApiClient.API_FINHUB_AI_SPOTLIGHT_PORTFOLIO)]
+    public async ValueTask<ActionResult<ApiResp<PortfolioAnalysis>>> SpotlightPortfolioPlan([FromRoute] string id)
+    {
+        var (authErrorResult, currentUser) = await VerifyAuthTokenAndCurrentUser();
+        if (authErrorResult != null)
+        {
+            // current auth token and signed-in user should all be valid
+            return authErrorResult;
+        }
+
+        var portfolioPlan = await GetPortfolioPlanIfOwnedByUser(currentUser, id);
+        if (portfolioPlan == null)
+        {
+            return ResponseNoData(404, "Portfolio plan not found.");
+        }
+
+        var portfolio = !string.IsNullOrEmpty(portfolioPlan.PortfolioId)
+            ? await PortfolioRepository.GetPortfolioByIdAsync(portfolioPlan.PortfolioId)
+            : null;
+        var market = Globals.MarketsMap.TryGetValue(portfolio?.Metadata?.DefaultMarketId?.ToUpper() ?? string.Empty, out var m) ? m : null;
+
+        // Step 0: if current holding is empty, failed immediately
+        var countPositiveAllocation = (portfolioPlan.Metadata?.HoldingTickers ?? []).Where(ht => ht.Shares > 0).Count();
+        if (countPositiveAllocation == 0)
+        {
+            return ResponseNoData(400, "Portfolio plan has no holdings, unable to provide spotlight analysis.");
+        }
+
+        // Step 1: make API call
+        var finhubResult = await SpotlightPortfolio(portfolioPlan, market);
+
+        // Step 2: build and return result
+        if (!finhubResult.IsSuccess)
+        {
+            return ResponseNoData(finhubResult.Status, finhubResult.Message ?? $"Failed to spotlight analyze portfolio plan '{portfolioPlan.Name}'", finhubResult.Extras);
+        }
+        var result = finhubResult.Data ?? new PortfolioAnalysis
+        {
+            LLMError = true,
+            LLMErrorMsg = "No data returned from FinHub API",
+            Analysis = string.Empty,
+        };
+        if (!result.LLMError)
+        {
+            portfolioPlan.Metadata ??= new();
+            portfolioPlan.Metadata.SpotlightRefreshTimestamp = DateTimeOffset.UtcNow.ToUnixTimeSeconds();
+            portfolioPlan.Metadata.Spotlight = result.Analysis;
+            await PortfolioRepository.UpdatePortfolioPlanAsync(portfolioPlan);
+        }
+        return ResponseOk(result);
+    }
+
+    private async ValueTask<ApiResp<PortfolioAnalysis>> SpotlightPortfolio(PortfolioPlanEntity plan, MarketDef? market)
+    {
+        Logger?.LogInformation("Calling FinHub SpotlightPortfolio API for portfolio plan '{planId}: {planName}' (market: {market}) with {numHoldings} holdings...", plan.Id, plan.Name, market?.Code ?? "N/A", plan.Metadata?.HoldingTickers?.Count ?? 0);
+        var req = new SpotLightPortfolioReq
+        {
+            Country = market?.Country ?? "US",
+            InvestorTheme = plan.Metadata?.Description,
+            CurrentAllocation = BuildHoldingTickersReq(plan)
+        };
+        return await FinHubClient.SpotlightPortfolioAsync(req);
     }
 
     /*----------------------------------------------------------------------*/